% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasim}
\alias{datasim}
\title{Simulated dataset for gamma family based PLSR}
\format{
A data frame with 200 observations on the following 8 variables.
\describe{ 
\item{Ygamma}{a numeric vector} 
\item{X1}{a numeric vector} 
\item{X2}{a numeric vector} 
\item{X3}{a numeric vector} 
\item{X4}{a numeric vector} 
\item{X5}{a numeric vector} 
\item{X6}{a numeric vector} 
\item{X7}{a numeric vector} 
\item{X8}{a numeric vector} 
}
}
\description{
This dataset provides a simulated dataset for gamma family based PLSR that was created with the \code{simul_data_UniYX_gamma} function.
}
\examples{

\donttest{
data(datasim)
X_datasim_train <- datasim[1:140,2:8]
y_datasim_train <- datasim[1:140,1]
X_datasim_test <- datasim[141:200,2:8]
y_datasim_test <- datasim[141:200,1]
rm(X_datasim_train,y_datasim_train,X_datasim_test,y_datasim_test)
}

}
\keyword{datasets}
