% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-bootPLS.R
\name{internal-bootPLS}
\alias{internal-bootPLS}
\alias{ust}
\alias{spls.dv}
\alias{correctp}
\alias{correctp.withoutK}
\alias{spls.Cboot}
\alias{cv.split}
\title{Internal bigPLS functions}
\description{
These are not to be called by the user.
}
\references{
A new bootstrap-based stopping criterion in PLS component construction,
J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand (2016), in The Multiple Facets of Partial Least Squares and Related Methods, 
\doi{10.1007/978-3-319-40643-5_18}\cr

A new universal resample-stable bootstrap-based stopping criterion for PLS component construction,
J. Magnanensi, F. Bertrand, M. Maumy-Bertrand and N. Meyer, (2017), Statistics and Compututing, 27, 757–774. 
\doi{10.1007/s11222-016-9651-4}\cr

New developments in Sparse PLS regression, J. Magnanensi, M. Maumy-Bertrand, 
N. Meyer and F. Bertrand, (2021), Frontiers in Applied Mathematics and Statistics, 
\doi{10.3389/fams.2021.693126}
}
\author{
Jérémy Magnanensi, Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{http://github.fbertran.io/homepage/}
}
\keyword{internal}
