% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{permcoefs.plsR.CSim}
\alias{permcoefs.plsR.CSim}
\title{Permutation bootstrap (Y,T) function for PLSR}
\usage{
permcoefs.plsR.CSim(dataset, i)
}
\arguments{
\item{dataset}{Dataset with tt}

\item{i}{Index for resampling}
}
\value{
Coefficient of the last variable in the linear regression 
\code{lm(dataset[i,1] ~ dataset[,-1] - 1)} computed using permutation 
resampling.
}
\description{
Permutation bootstrap (Y,T) function for PLSR
}
\examples{
set.seed(314)
xran=matrix(rnorm(150),30,5)
permcoefs.plsR.CSim(xran,sample(1:30))
}
\references{
A new bootstrap-based stopping criterion in PLS component construction,
J. Magnanensi, M. Maumy-Bertrand, N. Meyer and F. Bertrand (2016), in The Multiple Facets of Partial Least Squares and Related Methods, 
\doi{10.1007/978-3-319-40643-5_18}\cr

A new universal resample-stable bootstrap-based stopping criterion for PLS component construction,
J. Magnanensi, F. Bertrand, M. Maumy-Bertrand and N. Meyer, (2017), Statistics and Computing, 27, 757–774. 
\doi{10.1007/s11222-016-9651-4}\cr

New developments in Sparse PLS regression, J. Magnanensi, M. Maumy-Bertrand, 
N. Meyer and F. Bertrand, (2021), Frontiers in Applied Mathematics and Statistics, 
\doi{10.3389/fams.2021.693126}\cr.
}
\author{
Jérémy Magnanensi, Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{https://fbertran.github.io/homepage/}
}
